Numbers are float64, not int64

+++ influxdb.go
@@ -113,7 +113,7 @@ func (s *InfluxDBSink) decodeStat(stat StatResult) ([]ptFields, []ptTags, error)
 					} else {
 						// Ugly code to fix broken unsigned op_id from the API
 						if km == "op_id" {
-							if vm.(int64) == (2 ^ 32 - 1) {
+							if vm.(float64) == (2 ^ 32 - 1) {
 								vm = -1
 							}
 						}
@@ -135,7 +135,7 @@ func (s *InfluxDBSink) decodeStat(stat StatResult) ([]ptFields, []ptTags, error)
 			} else {
 				// Ugly code to fix broken unsigned op_id from the API
 				if km == "op_id" {
-					if vm.(int64) == (2 ^ 32 - 1) {
+					if vm.(float64) == (2 ^ 32 - 1) {
 						vm = -1
 					}
 				}
